+++ docs/test_examples.py
@@ -1,5 +1,6 @@
 import pytest
 import pathlib
+import warnings
 
 
 docs = pathlib.Path(__file__).parent
@@ -17,4 +18,6 @@ def test_docs(example_file):
     """Make sure all example files execute without error"""
     filename = str(example_file)
     with open(filename, "rb") as file:
-        exec(compile(file.read(), filename, "exec"), {})
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", UserWarning)
+            exec(compile(file.read(), filename, "exec"), {})