+++ changelog.md
@@ -10,6 +10,9 @@
   [#21](https://github.com/dean0x7d/pybinding/issues/21), and 
   [#23](https://github.com/dean0x7d/pybinding/issues/23).
 
+* Fixed compatibility with new versions of `matplotlib`: import error
+  ([#11](https://github.com/dean0x7d/pybinding/issues/11)) and various warnings.
+
 * Dropped support for Python 3.5. You must have Python 3.6 or newer to install this version.